privfair_fl/requirements.txt
numpy==1.14.5
scipy
tensorflow-gpu==1.10
Pillow
matplotlib

# required for mri
tqdm
scikit-learn

#### HE requirements ####
tenseal